% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_LiWaP.R
\name{calculate_consumptive_water}
\alias{calculate_consumptive_water}
\title{Calculate consumptive water use}
\usage{
calculate_consumptive_water(
  total_water_intake,
  feces_water_output,
  urine_water_output
)
}
\arguments{
\item{total_water_intake}{Numeric. Total daily water intake (L/head/day).}

\item{feces_water_output}{Numeric. Daily fecal water output (L/head/day).}

\item{urine_water_output}{Numeric. Daily urinary water output (L/head/day).}
}
\value{
Numeric. Daily consumptive water use (L/head/day).
}
\description{
Calculates daily consumptive water use by subtracting the recoverable share
of fecal and urinary water deposited on grazing land from total water intake.
}
\details{
Consumptive water use is calculated as:
\preformatted{
ConsumptiveWater =
WaterIntake - (WaterFeces + WaterUrine)
}
}
\seealso{
\code{\link{run_liwap}}
}
