% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiWaP-package.R
\docType{package}
\name{LiWaP-package}
\alias{LiWaP}
\alias{LiWaP-package}
\title{LiWaP: Livestock Water Partitioning Model}
\description{
Functions to estimate the partitioning of water in livestock production
systems at the cohort level. LiWaP calculates feed water, drinking water,
total water intake, water outputs, and consumptive water use using
species-specific coefficients.
}
\author{
\strong{Maintainer}: Lydia Lanzoni \email{lydia.lanzoni@fao.org}

Authors:
\itemize{
  \item Lydia Lanzoni \email{lydia.lanzoni@fao.org}
}

}
\keyword{internal}
